Remove test_all.dart, with the new test runner the file become unnecessary (is outdated anyway and drone.io build are already broken) But all tests into groups Create shared $vX shortcuts as a short version of creating a new vector (we hand such functions already previously, but every file defined own ones...) Split up test into own file per class
