+++ test/frustum_test.dart
@@ -10,27 +10,23 @@ import 'package:vector_math/vector_math.dart';
 
 import 'test_utils.dart';
 
-Vector3 _v3(double x, double y, double z) {
-  return new Vector3(x, y, z);
-}
-
 void testFrustumContainsVector3() {
   final Frustum frustum =
       new Frustum.matrix(makeFrustumMatrix(-1.0, 1.0, -1.0, 1.0, 1.0, 100.0));
 
-  expect(frustum.containsVector3(_v3(0.0, 0.0, 0.0)), equals(false));
-  expect(frustum.containsVector3(_v3(0.0, 0.0, -50.0)), equals(true));
-  expect(frustum.containsVector3(_v3(0.0, 0.0, -1.001)), equals(true));
-  expect(frustum.containsVector3(_v3(-1.0, -1.0, -1.001)), equals(true));
-  expect(frustum.containsVector3(_v3(-1.1, -1.1, -1.001)), equals(false));
-  expect(frustum.containsVector3(_v3(1.0, 1.0, -1.001)), equals(true));
-  expect(frustum.containsVector3(_v3(1.1, 1.1, -1.001)), equals(false));
-  expect(frustum.containsVector3(_v3(0.0, 0.0, -99.999)), equals(true));
-  expect(frustum.containsVector3(_v3(-99.999, -99.999, -99.999)), equals(true));
-  expect(frustum.containsVector3(_v3(-100.1, -100.1, -100.1)), equals(false));
-  expect(frustum.containsVector3(_v3(99.999, 99.999, -99.999)), equals(true));
-  expect(frustum.containsVector3(_v3(100.1, 100.1, -100.1)), equals(false));
-  expect(frustum.containsVector3(_v3(0.0, 0.0, -101.0)), equals(false));
+  expect(frustum.containsVector3($v3(0.0, 0.0, 0.0)), isFalse);
+  expect(frustum.containsVector3($v3(0.0, 0.0, -50.0)), isTrue);
+  expect(frustum.containsVector3($v3(0.0, 0.0, -1.001)), isTrue);
+  expect(frustum.containsVector3($v3(-1.0, -1.0, -1.001)), isTrue);
+  expect(frustum.containsVector3($v3(-1.1, -1.1, -1.001)), isFalse);
+  expect(frustum.containsVector3($v3(1.0, 1.0, -1.001)), isTrue);
+  expect(frustum.containsVector3($v3(1.1, 1.1, -1.001)), isFalse);
+  expect(frustum.containsVector3($v3(0.0, 0.0, -99.999)), isTrue);
+  expect(frustum.containsVector3($v3(-99.999, -99.999, -99.999)), isTrue);
+  expect(frustum.containsVector3($v3(-100.1, -100.1, -100.1)), isFalse);
+  expect(frustum.containsVector3($v3(99.999, 99.999, -99.999)), isTrue);
+  expect(frustum.containsVector3($v3(100.1, 100.1, -100.1)), isFalse);
+  expect(frustum.containsVector3($v3(0.0, 0.0, -101.0)), isFalse);
 }
 
 void testFrustumIntersectsWithSphere() {
@@ -38,50 +34,47 @@ void testFrustumIntersectsWithSphere() {
       new Frustum.matrix(makeFrustumMatrix(-1.0, 1.0, -1.0, 1.0, 1.0, 100.0));
 
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, 0.0), 0.0)), equals(false));
+      new Sphere.centerRadius($v3(0.0, 0.0, 0.0), 0.0)), isFalse);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, 0.0), 0.9)), equals(false));
+      new Sphere.centerRadius($v3(0.0, 0.0, 0.0), 0.9)), isFalse);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, 0.0), 1.1)), equals(true));
+      new Sphere.centerRadius($v3(0.0, 0.0, 0.0), 1.1)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, -50.0), 0.0)), equals(true));
+      new Sphere.centerRadius($v3(0.0, 0.0, -50.0), 0.0)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, -1.001), 0.0)), equals(true));
+      new Sphere.centerRadius($v3(0.0, 0.0, -1.001), 0.0)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(-1.0, -1.0, -1.001), 0.0)), equals(true));
+      new Sphere.centerRadius($v3(-1.0, -1.0, -1.001), 0.0)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(-1.1, -1.1, -1.001), 0.0)), equals(false));
+      new Sphere.centerRadius($v3(-1.1, -1.1, -1.001), 0.0)), isFalse);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(-1.1, -1.1, -1.001), 0.5)), equals(true));
+      new Sphere.centerRadius($v3(-1.1, -1.1, -1.001), 0.5)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(1.0, 1.0, -1.001), 0.0)), equals(true));
+      new Sphere.centerRadius($v3(1.0, 1.0, -1.001), 0.0)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(1.1, 1.1, -1.001), 0.0)), equals(false));
+      new Sphere.centerRadius($v3(1.1, 1.1, -1.001), 0.0)), isFalse);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(1.1, 1.1, -1.001), 0.5)), equals(true));
+      new Sphere.centerRadius($v3(1.1, 1.1, -1.001), 0.5)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, -99.999), 0.5)), equals(true));
+      new Sphere.centerRadius($v3(0.0, 0.0, -99.999), 0.5)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, -99.999), 0.0)), equals(true));
+      new Sphere.centerRadius($v3(0.0, 0.0, -99.999), 0.0)), isTrue);
   expect(frustum.intersectsWithSphere(
-          new Sphere.centerRadius(_v3(-99.999, -99.999, -99.999), 0.0)),
-      equals(true));
+      new Sphere.centerRadius($v3(-99.999, -99.999, -99.999), 0.0)), isTrue);
   expect(frustum.intersectsWithSphere(
-          new Sphere.centerRadius(_v3(-100.1, -100.1, -100.1), 0.0)),
-      equals(false));
+      new Sphere.centerRadius($v3(-100.1, -100.1, -100.1), 0.0)), isFalse);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(-100.1, -100.1, -100.1), 0.5)), equals(true));
+      new Sphere.centerRadius($v3(-100.1, -100.1, -100.1), 0.5)), isTrue);
   expect(frustum.intersectsWithSphere(
-          new Sphere.centerRadius(_v3(99.999, 99.999, -99.999), 0.0)),
-      equals(true));
+      new Sphere.centerRadius($v3(99.999, 99.999, -99.999), 0.0)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(100.1, 100.1, -100.1), 0.0)), equals(false));
+      new Sphere.centerRadius($v3(100.1, 100.1, -100.1), 0.0)), isFalse);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(100.1, 100.1, -100.1), 0.2)), equals(true));
+      new Sphere.centerRadius($v3(100.1, 100.1, -100.1), 0.2)), isTrue);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, -101.0), 0.0)), equals(false));
+      new Sphere.centerRadius($v3(0.0, 0.0, -101.0), 0.0)), isFalse);
   expect(frustum.intersectsWithSphere(
-      new Sphere.centerRadius(_v3(0.0, 0.0, -101.0), 1.1)), equals(true));
+      new Sphere.centerRadius($v3(0.0, 0.0, -101.0), 1.1)), isTrue);
 }
 
 void testFrustumIntersectsWithAabb3() {
@@ -89,23 +82,23 @@ void testFrustumIntersectsWithAabb3() {
       new Frustum.matrix(makeFrustumMatrix(-1.0, 1.0, -1.0, 1.0, 1.0, 100.0));
 
   expect(frustum.intersectsWithAabb3(new Aabb3.minMax(
-      v3(500.0, 500.0, 500.0), v3(1000.0, 1000.0, 1000.0))), isFalse);
+      $v3(500.0, 500.0, 500.0), $v3(1000.0, 1000.0, 1000.0))), isFalse);
   expect(frustum.intersectsWithAabb3(new Aabb3.minMax(
-      v3(-150.0, -150.0, -150.0), v3(150.0, 150.0, 150.0))), isTrue);
+      $v3(-150.0, -150.0, -150.0), $v3(150.0, 150.0, 150.0))), isTrue);
   expect(frustum.intersectsWithAabb3(
-      new Aabb3.minMax(v3(-1.5, -1.5, -1.5), v3(1.5, 1.5, 1.5))), isTrue);
+      new Aabb3.minMax($v3(-1.5, -1.5, -1.5), $v3(1.5, 1.5, 1.5))), isTrue);
   expect(frustum.intersectsWithAabb3(
-      new Aabb3.minMax(v3(0.0, 0.0, -50.0), v3(1.0, 1.0, -49.0))), isTrue);
+      new Aabb3.minMax($v3(0.0, 0.0, -50.0), $v3(1.0, 1.0, -49.0))), isTrue);
   expect(frustum.intersectsWithAabb3(
-      new Aabb3.minMax(v3(0.0, 0.0, 50.0), v3(1.0, 1.0, 51.0))), isFalse);
+      new Aabb3.minMax($v3(0.0, 0.0, 50.0), $v3(1.0, 1.0, 51.0))), isFalse);
   expect(frustum.intersectsWithAabb3(
-      new Aabb3.minMax(v3(0.0, 0.0, -0.99), v3(1.0, 1.0, 1.0))), isFalse);
+      new Aabb3.minMax($v3(0.0, 0.0, -0.99), $v3(1.0, 1.0, 1.0))), isFalse);
   expect(frustum.intersectsWithAabb3(
-      new Aabb3.minMax(v3(0.0, 0.0, -1.0), v3(1.0, 1.0, 1.0))), isTrue);
+      new Aabb3.minMax($v3(0.0, 0.0, -1.0), $v3(1.0, 1.0, 1.0))), isTrue);
   expect(frustum.intersectsWithAabb3(
-      new Aabb3.minMax(v3(0.0, 1.0, -10.0), v3(1.0, 2.0, 15.0))), isTrue);
+      new Aabb3.minMax($v3(0.0, 1.0, -10.0), $v3(1.0, 2.0, 15.0))), isTrue);
   expect(frustum.intersectsWithAabb3(
-      new Aabb3.minMax(v3(1.1, 1.1, -1.0), v3(2.0, 2.0, 0.0))), isFalse);
+      new Aabb3.minMax($v3(1.1, 1.1, -1.0), $v3(2.0, 2.0, 0.0))), isFalse);
 }
 
 void testFrustumCalculateCorners() {
