+++ test/sphere_test.dart
@@ -11,10 +11,10 @@ import 'package:vector_math/vector_math.dart';
 import 'test_utils.dart';
 
 void testSphereContainsVector3() {
-  final Sphere parent = new Sphere.centerRadius(v3(1.0, 1.0, 1.0), 2.0);
-  final Vector3 child = v3(1.0, 1.0, 2.0);
-  final Vector3 cutting = v3(1.0, 3.0, 1.0);
-  final Vector3 outside = v3(-10.0, 10.0, 10.0);
+  final parent = new Sphere.centerRadius($v3(1.0, 1.0, 1.0), 2.0);
+  final child = $v3(1.0, 1.0, 2.0);
+  final cutting = $v3(1.0, 3.0, 1.0);
+  final outside = $v3(-10.0, 10.0, 10.0);
 
   expect(parent.containsVector3(child), isTrue);
   expect(parent.containsVector3(cutting), isFalse);
@@ -22,10 +22,10 @@ void testSphereContainsVector3() {
 }
 
 void testSphereIntersectionVector3() {
-  final Sphere parent = new Sphere.centerRadius(v3(1.0, 1.0, 1.0), 2.0);
-  final Vector3 child = v3(1.0, 1.0, 2.0);
-  final Vector3 cutting = v3(1.0, 3.0, 1.0);
-  final Vector3 outside = v3(-10.0, 10.0, 10.0);
+  final parent = new Sphere.centerRadius($v3(1.0, 1.0, 1.0), 2.0);
+  final child = $v3(1.0, 1.0, 2.0);
+  final cutting = $v3(1.0, 3.0, 1.0);
+  final outside = $v3(-10.0, 10.0, 10.0);
 
   expect(parent.intersectsWithVector3(child), isTrue);
   expect(parent.intersectsWithVector3(cutting), isTrue);
@@ -33,10 +33,10 @@ void testSphereIntersectionVector3() {
 }
 
 void testSphereIntersectionSphere() {
-  final Sphere parent = new Sphere.centerRadius(v3(1.0, 1.0, 1.0), 2.0);
-  final Sphere child = new Sphere.centerRadius(v3(1.0, 1.0, 2.0), 1.0);
-  final Sphere cutting = new Sphere.centerRadius(v3(1.0, 6.0, 1.0), 3.0);
-  final Sphere outside = new Sphere.centerRadius(v3(10.0, -1.0, 1.0), 1.0);
+  final parent = new Sphere.centerRadius($v3(1.0, 1.0, 1.0), 2.0);
+  final child = new Sphere.centerRadius($v3(1.0, 1.0, 2.0), 1.0);
+  final cutting = new Sphere.centerRadius($v3(1.0, 6.0, 1.0), 3.0);
+  final outside = new Sphere.centerRadius($v3(10.0, -1.0, 1.0), 1.0);
 
   expect(parent.intersectsWithSphere(child), isTrue);
   expect(parent.intersectsWithSphere(cutting), isTrue);
@@ -44,7 +44,9 @@ void testSphereIntersectionSphere() {
 }
 
 void main() {
-  test('Sphere Contains Vector3', testSphereContainsVector3);
-  test('Sphere Intersection Vector3', testSphereIntersectionVector3);
-  test('Sphere Intersection Sphere', testSphereIntersectionSphere);
+  group('Sphere', () {
+    test(' Contains Vector3', testSphereContainsVector3);
+    test('Sphere Intersection Vector3', testSphereIntersectionVector3);
+    test('Sphere Intersection Sphere', testSphereIntersectionSphere);
+  });
 }
